refactor: user activity file naming
